+++ src/lib/components/Tools/ZoomTool.svelte
